services/extract-archive/requirements.txt
aiofiles==23.2.1
aio-pika==9.5.4
cuid2==2.0.1
patool==3.1.0
aiohttp==3.11.11
minio==7.2.15
